Add alt tag generator example (#471)
Added a alt tag generator that will generate relevant alt text for
images using an ML model. First time contributing to this repo so please
let me know if I missed anything or if there'd be a better folder for
this!

I'm going to be adding this to templates marketplace after this is
merged, and sharing it on Twitter.