Update aws-s3-image-upload to Next 13.2
During my own development, I modified the existing aws-s3-image-upload example to use Next.js 13.2's new /app/ directory and updated the API to use the new route handlers.

This is a stripped down and re-styled version of a project I was working on that I would like to contribute to open-source to make it easier for new users to adopt Next.js with AWS S3 SDK.

I have included a long list of resources and documentation I used throughout this in the README.

[Changes]
- Improved clarity of the previous README by adding steps in between on where to navigate on AWS.
- Removed README instructions for CDK. This is more beginner friendly and outlines all of the same steps with more clarity by using the AWS Console.
- Removed bucket policies as it is redundant to create an AWS IAM with FullS3Access, and also to specify granular permissions.
- Added relevant resources for further understanding of how AWS SDK / Next.js 13 operates with common answers to questions I had during development.
- Added correct CORS policies.
- Changed S3Client to include region, as per the example in the AWS documentation.
- Updated styling on the example to look more visually appealing.
- Added NEXT_PUBLIC_BASE_URL to direct the API route, which can be adjusted to work on either development or deployment.

[Re: Issue #609]
User @kdesimini mentioned the previous guide was confusing and outdated. I ran into a similar experience following the previous guide.