bug: Instantiating the wasm directly cause incorrect behaviour
When instantiation the wasm with WebAssembly.instantiate(module), if the
rust application required any usage of the js stack, the route would
failed due to the lack of pointer to any imports required by
wasm-bindgen.

Using the wasm-pack target web, the generated code comes with a more
suitable loader for the webassembler which will load all the necessary
imports in the correct place.