Update Flask example 2.2.2 to 2.2.5, add Flask 3.0.0 example (#798)
### Description

The Flask 2.2.2 example relies on `Werkzeug >= 2.2.2` which today
installs `Werkzeug 2.3.7` and doesn't appear to be compatible with Flask
2.2.3 and older. The Flask example needs to be updated to 2.2.4 or
newer. Opted to update to latest 2.2.5.

This PR renames the `flask` example to `flask2` now that Flask 3 is
available (released Sept 30, 2023), then adds a Flask 3 example. If
preferred, happy to remove the `flask3`, rename `flask2` back to `flask`
and update it from 2.2.2 straight to 3.0.0. My only concern is Flask 3
is pretty new and there could be incompatibilities beyond our control
with dependencies. I did test the example with 3.0.0 and it worked.

Please TEST this before merging. Thanks!

### Demo URL

<!--
Provide a URL to a live deployment where we can test your PR. If a demo
isn't possible feel free to omit this section.
-->

### Type of Change

- [x] New Example
- [x] Example updates (Bug fixes, new features, etc.)
- [ ] Other (changes to the codebase, but not to examples)

### New Example Checklist

- [ ] 🛫 `npm run new-example` was used to create the example
- [ ] 📚 The template wasn't used but I carefuly read the [Adding a new
example](https://github.com/vercel/examples#adding-a-new-example) steps
and implemented them in the example
- [ ] 📱 Is it responsive? Are mobile and tablets considered?