Remove numpy from nextjs-flask
pip fails to install this dependency on Python 3.12 https://github.com/vercel/examples/issues/946

While fixing the error is certainly one way of going about it, numpy is not even used for the hello-world Flask app this ships with. To simplify things, this commit removes numpy as a dependency entirely.