[microfrontends] Add Next.js Multi-Zones example using Vercel Microfrontends (#1259)
### Description

This PR adds a basic Next.js Multi-Zones example using Vercel
Microfrontends to the `vercel/examples` repo. This is a copy of the
current template
(https://vercel.com/templates/microfrontends/microfrontends-multi-zones)
moved to `vercel/examples` with some clean-ups to the README.md.

### Demo URL

<!--
Provide a URL to a live deployment where we can test your PR. If a demo
isn't possible feel free to omit this section.
-->

### Type of Change

- [ ] New Example
- [ ] Example updates (Bug fixes, new features, etc.)
- [ ] Other (changes to the codebase, but not to examples)

### New Example Checklist

- [ ] 🛫 `npm run new-example` was used to create the example
- [ ] 📚 The template wasn't used but I carefuly read the [Adding a new
example](https://github.com/vercel/examples#adding-a-new-example) steps
and implemented them in the example
- [ ] 📱 Is it responsive? Are mobile and tablets considered?