[cdn] Create a new template for creating a docs rewrite to mintlify (#1362)
### Description

This is a template for using a rewrites in vercel.ts to a docs site.
This is meant to display the capabilities of the new vercel.ts project.,
a typsecript based configuration file. Please view the readme for more
info.

### Demo URL

https://docsrewrite.vercel.app/

### Type of Change

- [x] New Example
- [ ] Example updates (Bug fixes, new features, etc.)
- [ ] Other (changes to the codebase, but not to examples)

### New Example Checklist

- [ ] 🛫 `npm run new-example` was used to create the example
- [ ] 📚 The template wasn't used but I carefuly read the [Adding a new
example](https://github.com/vercel/examples#adding-a-new-example) steps
and implemented them in the example
- [ ] 📱 Is it responsive? Are mobile and tablets considered?