[cdn] Create a new template for creating bulk redirects via a CMS (#1336)
### Description

This is a template for using a CMS with bulk redirects to auto update
the redirects at build time based on new entries and updates on the CMS.
In this case the CMS is Contentful. This is meant to display our new
typsecript based configuration file. Please view the readme for more
info.

### Demo URL

https://cms-bulk-redirects.vercel.app/

### Type of Change

- [x] New Example
- [ ] Example updates (Bug fixes, new features, etc.)
- [ ] Other (changes to the codebase, but not to examples)

### New Example Checklist

- [ ] 🛫 `npm run new-example` was used to create the example
- [x] 📚 The template wasn't used but I carefuly read the [Adding a new
example](https://github.com/vercel/examples#adding-a-new-example) steps
and implemented them in the example
- [x] 📱 Is it responsive? Are mobile and tablets considered?